SWDEV-548892 - Partially avoid using ockl workitem ID functions
The same functions are already directly implemented in clang. The development
history is messy. The builtins should implement both the COv4 or COv5+ ABI
depending on the compiler flag, but we somehow ended up with versioned
ABI paths in both the OCKL functions which are also partially implemented
with these builtins. It's simplest to just use the builtin.

This also drops the handling of the global offset from the
__hip_get_block_dim_{x|y|z} functions, and hc_get_workitem_absolute_id.
This is an OpenCL feature and I do not think is accessible from HIP, so
this is theoretically faster.

I'm also not sure if there's any reason for hc_get_workitem_absolute_id
to still exist. There are no references to it in the repo, besides
this definition.

Leaves __ockl_get_num_groups alone for now. The replacement should be
`__builtin_amdgcn_grid_size_x() / __builtin_amdgcn_workgroup_size_x()`.
However, that needs to be accompanied by an optimization to turn that into
a load from the new ABI location which apparently hasn't been implemented
yet.